requirements.txt
python-telegram-bot>=20.0
playwright
websockets
cryptography
python-dotenv
fastapi
uvicorn[standard]
httpx
yfinance>=0.2.40
